+++ src/bindings/Factory.ts
@@ -10,6 +10,6 @@ export class Factory extends ResolverBinding implements FactoryInterface {
    * @return {any}
    */
   public resolve (container: ContainerInterface): any {
-    return this._resolver(container)
+    return this.resolver(container)
   }
 }