+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Removed
 - `getBindings()` method from `Container` class
+- `getFakes()` method from `Container` class
 
 ## [0.4.0]
 ### Added