+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [1.0.0]
 ### Removed
 - `getBindings()` method from `Container` class
 - `getFakes()` method from `Container` class
@@ -66,7 +68,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Binding` class with `BindingInterface`
 - `Container` class with `ContainerInterface`
 
-[Unreleased]: https://github.com/halliganjs/service-container/compare/v0.4.0...HEAD
+[Unreleased]: https://github.com/halliganjs/service-container/compare/v1.0.0...HEAD
+[1.0.0]: https://github.com/halliganjs/service-container/compare/v0.4.0...v1.0.0
 [0.4.0]: https://github.com/halliganjs/service-container/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/halliganjs/service-container/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/halliganjs/service-container/compare/v0.2.0...v0.2.1